+++ frontend/guest-app/src/components/Question/QuestionContainer.js
@@ -14,22 +14,13 @@ import {GuestGlobalContext} from "../../libs/guestGlobalContext.js";
 import {QuestionsProvider} from "./QuestionsContext.js";
 import PaddingArea from "./QuestionInputArea/PaddingArea.js";
 import {ContainerProvider} from "./ContainerContext.js";
-import QuestionInputDrawer from "./QuestionInputArea/QuestionInputDrawer.js";
 import useToggleReducer from "./useToggleReducer.js";
 import QuestionEditMenuDrawer from "./QuestionCard/QuestionEditMenuDrawer.js";
+import NewQuestionInputDrawer from "./NewQuestionInputDrawer.js";
+import EditQuestionInputDrawer from "./EditQuestionInputDrawer.js";
 
 const RECENT_TAB_IDX = 1;
 const POPULAR_TAB_IDX = 2;
-
-function getNewQuestion({EventId, GuestId, guestName, content}) {
-	return {
-		guestName,
-		EventId,
-		GuestId,
-		content,
-		isAnonymous: guestName.length === 0,
-	};
-}
 
 const useDataLoadEffect = (dispatch, data) => {
 	useEffect(() => {
@@ -72,7 +63,6 @@ const useSocketHandler = (dispatch, guestGlobal) => {
 		dispatch({type: "removeQuestion", data: req});
 	});
 
-
 	useSocket("question/update", req => {
 		req.guestGlobal = guestGlobal;
 		dispatch({type: "updateQuestion", data: req});
@@ -96,18 +86,6 @@ function QuestionContainer() {
 	useDataLoadEffect(dispatch, data);
 	useSocketHandler(dispatch, guest);
 
-	const onConfirmNewQuestion = () => {
-		socketClient.emit(
-			"question/create",
-			getNewQuestion({
-				guestName: userNameRef.current.value,
-				EventId: event.id,
-				GuestId: guest.id,
-				content: questionRef.current.value,
-			}),
-		);
-	};
-
 	const onContainerSelectTab = (event, newValue) => {
 		if (newValue === RECENT_TAB_IDX) {
 			dispatch({type: "sortByRecent"});
@@ -118,40 +96,6 @@ function QuestionContainer() {
 		}
 
 		selectTabIdx(event, newValue);
-	};
-
-	const newQuestionInputDrawerProps = {
-		title: "질문 하기",
-		isOpen: newQuestionInputDrawerReducer.state,
-		onClose: () => newQuestionInputDrawerReducer.setOff(),
-		onConfirm: onConfirmNewQuestion,
-		userNameRef,
-		questionRef,
-	};
-
-	const onConfirmEditQuestion = () => {
-		socketClient.emit("question/update", {
-			id: EditQuestionInputDrawerReducer.data.id,
-			guestName: userNameRef.current.value,
-			EventId: event.id,
-			GuestId: guest.id,
-			content: questionRef.current.value,
-			isAnonymous: userNameRef.current.value.length === 0,
-		});
-	};
-
-	const editQuestionInputDrawerProps = {
-		title: "질문 수정 하기",
-		isOpen: EditQuestionInputDrawerReducer.state,
-		onClose: () =>
-			EditQuestionInputDrawerReducer.setOff(
-				EditQuestionInputDrawerReducer.data,
-			),
-		onConfirm: onConfirmEditQuestion,
-		userNameRef,
-		questionRef,
-		initialUserName: EditQuestionInputDrawerReducer.data.guestName,
-		initialQuestion: EditQuestionInputDrawerReducer.data.content,
 	};
 
 	return (
@@ -171,8 +115,16 @@ function QuestionContainer() {
 				<AddQuestionInputButton
 					onClick={() => newQuestionInputDrawerReducer.setOn()}
 				/>
-				<QuestionInputDrawer {...newQuestionInputDrawerProps} />
-				<QuestionInputDrawer {...editQuestionInputDrawerProps} />
+				<NewQuestionInputDrawer
+					userNameRef={userNameRef}
+					questionRef={questionRef}
+					toggleReducer={newQuestionInputDrawerReducer}
+				/>
+				<EditQuestionInputDrawer
+					userNameRef={userNameRef}
+					questionRef={questionRef}
+					toggleReducer={EditQuestionInputDrawerReducer}
+				/>
 				<QuestionEditMenuDrawer
 					isOpen={questionEditMenuReducer.state}
 					onClose={() => questionEditMenuReducer.setOff()}
